#77 improve documentation; move class and interfaces to separate files
