+++ src/main/kotlin/io/kotlintest/InvalidConfigException.kt
@@ -1,0 +1,6 @@
+package io.kotlintest
+
+/**
+ * Signals an invalid configuration
+ */
+class InvalidConfigException(message: String) : Exception(message)
