+++ src/main/kotlin/io/kotlintest/TestCaseContext.kt
@@ -1,0 +1,8 @@
+package io.kotlintest
+
+/**
+ * Meta information about a [TestCase].
+ */
+data class TestCaseContext(
+    val spec: TestBase,
+    val testCase: TestCase)
